first unit test now works

+++ src/build.sh
@@ -1,14 +1,21 @@
 #!/bin/bash
 
-rm -rf build
-mkdir -p build
+CMD="$1"
+
+if [ "X$1" == "Xrebuild" ]; then
+    rm -rf build
+    mkdir -p build
+    (
+        cd build
+        cmake ..
+        if [ $? -ne 0 ]; then
+            echo "cmake failed"
+            exit 1
+        fi
+    )
+fi
 
 (
     cd build
-    cmake ..
-    if [ $? -ne 0 ]; then
-        echo "cmake failed"
-        exit 1
-    fi
     make
 )
